M7: запись в теги — batch-контракт, Modbus FC05/06/16, очередь в канал, аудит SQLite, персистентность internal-тегов; фикс точности RegisterDecoder для 32-битных типов

+++ SCADA.Drivers.Modbus/RegisterDecoder.cs
@@ -10,13 +10,16 @@ public static class RegisterDecoder
     /// <summary>data — блок ответа, byteOffset — смещение тега в БАЙТАХ (регистры × 2).</summary>
     public static double Decode(ReadOnlySpan<byte> data, int byteOffset, ModbusDataType type)
     {
+        // Все ветки приводятся к double ЯВНО: без этого общий тип switch-выражения
+        // выводится как float (из-за ветки Float32), и uint/int больше 2^24
+        // теряют точность ещё до приведения результата.
         return type switch
         {
-            ModbusDataType.UInt16 => (data[byteOffset] << 8) | data[byteOffset + 1],
-            ModbusDataType.Int16 => (short)((data[byteOffset] << 8) | data[byteOffset + 1]),
-            ModbusDataType.UInt32 => ReadUInt32BigEndian(data, byteOffset),
-            ModbusDataType.Int32 => (int)ReadUInt32BigEndian(data, byteOffset),
-            ModbusDataType.Float32 => BitConverter.Int32BitsToSingle(
+            ModbusDataType.UInt16 => (double)((data[byteOffset] << 8) | data[byteOffset + 1]),
+            ModbusDataType.Int16 => (double)(short)((data[byteOffset] << 8) | data[byteOffset + 1]),
+            ModbusDataType.UInt32 => (double)ReadUInt32BigEndian(data, byteOffset),
+            ModbusDataType.Int32 => (double)(int)ReadUInt32BigEndian(data, byteOffset),
+            ModbusDataType.Float32 => (double)BitConverter.Int32BitsToSingle(
                 (int)ReadUInt32BigEndian(data, byteOffset)),
             _ => throw new ArgumentOutOfRangeException(nameof(type), type, null)
         };
